% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R, R/utils.R
\name{download_metadata}
\alias{download_metadata}
\title{Support function to download metadata internally used}
\usage{
download_metadata()

download_metadata()
}
\description{
Support function to download metadata internally used

Support function to download metadata internally used
Funcion para descargar la lista de datos espaciales de peru
}
\examples{
\dontrun{ if (interactive()) {
df <- download_metadata()
}}
\dontrun{ if (interactive()) {
df <- download_metadata()
}}
}
\keyword{internal}
