% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_metadata_anp}
\alias{download_metadata_anp}
\title{Support function to download metadata internally used for anp data}
\usage{
download_metadata_anp()
}
\description{
Support function to download metadata internally used for anp data
}
\keyword{internal}
