% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anp_peru.R
\name{get_anp_peru}
\alias{get_anp_peru}
\title{Download Spatial Data of Protected Natural Areas (ANP) in Peru}
\usage{
get_anp_peru(anp = NULL, showProgress = TRUE)
}
\arguments{
\item{anp}{A character or a vector with the name(s) of the protected natural areas in Peru.}

\item{showProgress}{Logical TRUE or FALSE to display a progress bar during download.}
}
\value{
An \verb{"sf" "data.frame"} object containing the spatial data of Peru's protected natural areas.
}
\description{
Downloads spatial data of protected natural areas in Peru declared by SERNANP
(National Service of Natural Areas Protected by the State).
Data were obtained from the
\href{https://www.gob.pe/institucion/sernanp/pages/21261-acceder-a-informacion-espacial-de-las-area-naturales-protegidas-visor-de-informacion-geografica}{SERNANP Geographic Information Viewer}
as the official source. The data use WGS 84 (EPSG:4326).
}
\examples{
if (interactive()) {
# Read specific ANP
manu <- get_anp_peru(anp = "Manu")

}
}
