% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dep.R
\name{get_dep_sf}
\alias{get_dep_sf}
\title{Obtener datos espaciales de distritos por departamento en Perú}
\usage{
get_dep_sf(dep_name = NULL)
}
\arguments{
\item{dep_name}{El nombre o nombres de los departamentps a las que se desean filtrar los datos.
Puede ser un vector de cadenas de caracteres. Si se omite o es NULL, la función devolverá todos los distritos.}
}
\value{
Un objeto espacial (SpatialDataFrame) que contiene los datos de los distritos
de las provincias especificadas. Si no se encuentra ninguna coincidencia o los datos de Perú
no están disponibles, devuelve NULL.
}
\description{
Esta función permite filtrar los datos espaciales de los distritos de Perú
según el nombre o nombres de los departamentos especificados.
}
\examples{
# Obtener datos del departamento de Lima
lima_data <- get_dep_sf(c("LIMA"))

# Obtener datos de los departamentos

varios_data <- get_dep_sf(c("LIMA", "CUSCO"))

}
