% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distritos.R
\docType{data}
\name{peru}
\alias{peru}
\title{Districts of Peru}
\format{
An \code{sf} object with 1,874 rows and four variables:
\describe{
\item{departamento}{Department name.}
\item{provincia}{Province name.}
\item{distrito}{District name.}
\item{geometry}{Polygon geometry.}
}
}
\source{
\url{https://ide.inei.gob.pe/}
}
\usage{
peru
}
\description{
An \code{sf} object containing district-level administrative boundaries of Peru
in the WGS 84 coordinate reference system (EPSG:4326).
}
\keyword{dataset}
