% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_data_level}
\alias{select_data_level}
\title{Select data level: nacional, departamental, provincial}
\usage{
select_data_level(temp_meta, level = "prov")
}
\arguments{
\item{temp_meta}{A dataframe with the file_url addresses of perugpkg datasets}

\item{level}{'all', 'dep' or 'prov'}
}
\description{
Select data level: nacional, departamental, provincial
}
\keyword{internal}
