% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_metadata_anp}
\alias{select_metadata_anp}
\title{Select metadata of anp}
\usage{
select_metadata_anp(anp)
}
\arguments{
\item{anp}{Which anp will be downloaded.}
}
\description{
Select metadata of anp
}
\keyword{internal}
