% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prov.R
\name{get_prov_sf}
\alias{get_prov_sf}
\title{Obtener datos espaciales de distritos por provincia en Peru}
\usage{
get_prov_sf(prov_name = NULL)
}
\arguments{
\item{prov_name}{El nombre de la provincia a la que se desean filtrar los datos.
Debe ser una cadena de caracteres. Si se omite o es NULL, la función devolvera
todos los distritos.}
}
\value{
Un objeto espacial (SpatialDataFrame) que contiene los datos de los
distritos de la provincia especificada. Si no se encuentra ninguna coincidencia
devuelve NULL.
}
\description{
Esta funcion permite filtrar los datos espaciales de los distritos de Peru
según el nombre de la provincia especificada.
}
\examples{
# Obtener datos de la provincia de Lima
lima_data <- get_prov_sf("Lima")
# Obtener todos los datos de distritos sin filtrar por provincia
all_data <- get_prov_sf()
}
