% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distritos.R
\docType{data}
\name{peru}
\alias{peru}
\title{Distritos Peru}
\format{
A \code{SpatialPolygonsDataFrame}
}
\source{
\url{https://ide.inei.gob.pe/}
}
\usage{
peru
}
\description{
Información geográfica vectorial de los 1722 distritos del PERU y su geometría.
}
\keyword{dataset}
