% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{select_data_level}
\alias{select_data_level}
\title{Select data level: 'dep', 'prov' (default)}
\usage{
select_data_level(temp_meta, simplified = NULL)
}
\arguments{
\item{temp_meta}{A dataframe with the file_url addresses of geobr datasets}

\item{simplified}{Logical TRUE or FALSE indicating  whether the function
returns the 'complete' dataset or a dataset 'simplified' hiegher level (Defaults to TRUE)}
}
\description{
Select data level: 'dep', 'prov' (default)
}
\keyword{internal}
